Dockerfiles: add http_proxy settings
When installing KernelCI behind an http proxy we need
to set the http(s)_proxy variables. HTTP_PROXY is a predefined
build variable, so no need to specify it. We also set ENV
because it is needed by PIP, wget and others.
Note that ENV variables will also be present during runtime.

Signed-off-by: Daniel Sangorrin <daniel.sangorrin@toshiba.co.jp>